pyproject.toml
[tool.poetry]
name = "cociplan"
version = "1.6.0"
description = "Weekly menu planner"
authors = ["Rene Fernandez <renefernandez@duck.com>"]
license = "GNU General Public License v3 (GPLv3)"
readme = "README.md"
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2"
django-cors-headers = "3.14.0"
django-filter = "23.2"
djangorestframework = "3.14.0"
djangorestframework-camel-case = "1.4.2"
drf-extra-fields = "3.4.1"
drf-yasg = "1.21.5"
mysqlclient = "2.1.1"
pillow = "9.5.0"
sentry-sdk = "1.21.1"
django-environ = "^0.10.0"
python-decouple = "^3.8"
gunicorn = "^21.2.0"


[tool.poetry.group.dev.dependencies]
bandit = "1.7.4"
black = "23.1.0"
coverage = "6.2"
django-types = "0.17.0"
djangorestframework-stubs = "1.8.0"
djlint = "1.34.1"
factory-boy = "3.2.1"
faker = "11.3.0"
flake8 = "5.0.4"
flake8-django = "1.1.5"
isort = "5.12.0"
pyright = "1.1.296"
pre-commit = "2.15.0"
responses = "0.17.0"
types-requests = "2.28.11.13"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | client
  | docs
  | data
  | media
  | migrations
)/
'''

[tool.isort]
# make it compatible with black
profile = "black"
# isort configuration:
# https://github.com/PyCQA/isort/wiki/isort-Settings
include_trailing_comma = true
use_parentheses = true
# See https://github.com/PyCQA/isort#multi-line-output-modes
multi_line_output = 3
combine_as_imports = true
src_paths = ["backend"]

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "cociplan.settings"
python_files = ["tests.py", "test_*.py", "*_tests.py"]
pythonpath = [".", "backend"]
